% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mntcapp.R
\name{mntcapp}
\alias{mntcapp}
\title{Bootstrap for App for Form Generation
A generic inspection form (not technically Amazon affiliated)}
\usage{
mntcapp()
}
\value{
None
}
\description{
Bootstrap for App for Form Generation
A generic inspection form (not technically Amazon affiliated)
}
\examples{
\dontrun{

##importFrom shinyBS
##importFrom ggplot2
}
print("this has not yet been scripted")
}
\author{
Samuel Mark DeClerk
}
