% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunction.R
\name{testFunction}
\alias{testFunction}
\title{Testing dodcumentation on this package
This is a program that says your vagina is nasty 9999 times}
\usage{
testFunction(name)
}
\arguments{
\item{name}{an object of type string that takes a name and completes a sentence about how nasty your vagina is. If blank, name is substituted for the current username.}
}
\value{
string_n
}
\description{
Testing dodcumentation on this package
This is a program that says your vagina is nasty 9999 times
}
\examples{
V <- testFunction("Samuel")
V
V <- testFunction("")
V
}
\author{
Samuel MDC
}
